Add LocalStack validation pipeline (#11)
# Motivation

A release of this library has at least once broken the LocalStack CI
pipeline, with a spooky action at a distance. This is because importing
the JS libraries into running AppSync resolvers was broken, and so the
AppSync JavaScript resolver tests failed.

# Changes

* Add a CI pipeline step that runs LocalStack, configuring the appsync
resolver JS library version to use the commit specified in the current
branch.

Closes #6