Use non-dir import for dynamodb
When importing the modules using the `scripts/coverage.mjs` script I got
an import error because the dynamodb utils use the main utils, and were
importing them by directory.

I do not understand JS import rules enough to debug this properly, so we
just import from the `index.js` instead. Let's see if the CI passes and
we can actually use the functions.