Capture function coverage (#20)
# Motivation

We should be transparent about which functions are covered and which
aren't. This helps our users understand our current status, and indicate
that we are not 100% complete with our implementation.

To aid with this process, I have created
https://github.com/localstack/appsync-utils/issues/19 which lists the
functions we support, and encourage users to list/upvote their
suggestions.
In turn, to support this I have added a script to capture the functions
we implement, and to print them out to the console. We may automate this
procedure in the process if it gets too arduous.

The example output from this script can be seen in the ["Currently
implemented
functions"](https://github.com/localstack/appsync-utils/issues/19#user-content-#currently-implemented-functions)
section of the linked issue.

# Changes

* Add script to capture feature coverage by introspecting the utilities
library.
* Fix a relative import issue when using the script