Fix `rds.toJsonObject` method only supporting string values
Currently, the `toJsonObject` method from `rds` only supports `{ stringValue: string }`, but other value types can appear in payloads, such as `{ longValue: int }`.

This PR changes it so it assumes value types of any kind, as long as the value type is an object with a key and the value we want to map.